+++ packages/agtree/CHANGELOG.md
@@ -5,12 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog][keepachangelog], and this project adheres to [Semantic Versioning][semver].
 
 
-## Unreleased
+## 1.1.5 - 2023-09-05
 
 ### Changed
 
 - Validation of `$csp` and `$permissions` modifiers value
   by custom pre-defined validator instead of regular expression
+
+### Added
+
+- Exports to `package.json`
 
 ## 1.1.4 - 2023-08-30